Test tags example
Let's move `tags` test script from selftests to examples. With such
change we can use this test in other places, for example in tags
documentation.
